services/image_processor.py
import cv2
import easyocr
import numpy as np
from typing import Dict, List, Tuple
import os
import json

class ImageProcessor:
    def __init__(self):
        self.reader = easyocr.Reader(['ch_sim', 'en'])
        self.pet_keywords = {
            'type': ['猫', '狗', '兔子'],
            'color': ['黑色', '白色', '橙色', '灰色', '棕色'],
            'features': ['项圈', '断尾', '断耳', '绝育']
        }

    def process_image(self, image_path: str) -> Dict:
        """处理图像并返回识别结果"""
        # 读取图像
        image = cv2.imread(image_path)
        if image is None:
            raise ValueError("无法读取图像")

        # OCR文本识别
        text_results = self.extract_text(image)
        
        # 特征识别
        features = self.detect_features(image)
        
        # 合并结果
        return {
            'text_content': text_results,
            'detected_features': features,
            'confidence_score': self.calculate_confidence(text_results, features)
        }

    def extract_text(self, image: np.ndarray) -> List[str]:
        """提取图像中的文字"""
        results = self.reader.readtext(image)
        return [text for _, text, conf in results if conf > 0.5]

    def detect_features(self, image: np.ndarray) -> Dict:
        """检测宠物特征"""
        # 这里将来可以集成更复杂的特征检测算法
        features = {
            'pet_type': None,
            'color': [],
            'special_features': []
        }
        return features

    def calculate_confidence(self, text_results: List[str], features: Dict) -> float:
        """计算识别结果的置信度"""
        # 基础实现，后续可以优化
        return 0.7

    def match_posts(self, post1_features: Dict, post2_features: Dict) -> float:
        """计算两个帖子之间的匹配度"""
        # 基础实现，后续可以添加更复杂的匹配算法
        score = 0.0
        # 实现匹配逻辑
        return score

def create_processor():
    """创建图像处理器实例"""
    return ImageProcessor()
