requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
wechatpy==1.8.18
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
geopy==2.4.0
opencv-python==4.8.0
easyocr==1.7.1
numpy==1.24.3
torch==2.0.1
gunicorn==21.2.0
gevent==23.9.1
supervisor==4.2.5
